+++ admin/category/list.php
@@ -34,8 +34,8 @@
 				  		//      <input type="checkbox" name="selector[]" id="selector[]" value="'.$result->CATEGORYID. '"/>
 				  		// 		' . $result->CATEGORIES.'</a></td>';
 				  			echo '<td>' . $result->CATEGORY.'</td>';
-				  		echo '<td align="center"><a title="Edit" href="index.php?view=edit&id='.$result->CATEGORYID.'" class="btn btn-primary btn-xs  ">  <span class="fa fa-edit fw-fa"></a>
-				  		     <a title="Delete" href="controller.php?action=delete&id='.$result->CATEGORYID.'" class="btn btn-danger btn-xs  ">  <span class="fa  fa-trash-o fw-fa "></a></td>';
+				  		echo '<td align="center"><a title="Edit" href="index.php?view=edit&id='.$result->CATEGORYID.'" class="btn btn-primary btn-xs  ">  <span class="fa fa-edit fw-fa"></span>Edit</a>
+								<a title="Delete" href="controller.php?action=delete&id='.$result->CATEGORYID.'" class="btn btn-danger btn-xs" onclick="return confirmDelete();">  <span class="fa  fa-trash-o fw-fa "></span></a></td>';
 				  		// echo '<td></td>';
 				  		echo '</tr>';
 				  	} 
@@ -54,4 +54,9 @@
 			
 				</form>
 	
+	<script>
+		function confirmDelete() {
+    	return confirm("Are you sure you want to delete this category?");
+		}
+	</script>
  <div class="table-responsive">
